+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Changelog
 All changes to this project will be documented in this file.
 
+## [0.0.5] - 2021-08-12
+- Fix chunk size (default 50MB, min 5MB, max 128MB)
+- Fix upload progress listener